backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic-core==2.33.1
sqlalchemy==2.0.23
aiofiles==23.2.1
python-multipart==0.0.6
httpx==0.25.1
requests==2.31.0
pycryptodome==3.22.0  # Alternative to pycrypto
PyPDF2==3.0.1
reportlab==4.0.7
python-dotenv==1.0.0
cohere==5.1.8
langchain-cohere==0.1.1
langchain-core==0.1.32

# Dependencies required for compatibility
anyio==3.7.1
starlette==0.27.0
typing-extensions==4.12.0
h11==0.14.0
certifi==2023.5.7
httpcore==0.18.0
greenlet==2.0.2
annotated-types==0.5.0
charset-normalizer==3.1.0
urllib3==2.0.3
langsmith==0.1.38
jsonpatch==1.33
packaging==23.1
pyyaml==6.0
tenacity==8.2.2
typing-inspection==0.4.0
orjson==3.9.5
sniffio==1.3.0
colorama==0.4.6 